♻️ move terminal component to its own component
